Changes to be committed: 	new file:   MoTF/Optimizer/Search.py 	new file:   MoTF/Optimizer/__init__.py 	modified:   MoTF/test.py
Search strategy has been implemented for global optimizer. It utilizes a decorator class to stop the search process in case of maximum number of iteration limit is exceeded. The thrown exception includes the valid results in the attribute object.

In global optimizer class, the instances has to be modified as command design pattern to call the same function of the objects for various domains.